+++ packages/common/src/main/kotlin/org/vechain/indexer/b3tr/action/repository/AppRoundActionSummaryRepository.kt
@@ -1,5 +1,6 @@
 package org.vechain.indexer.b3tr.action.repository
 
+import java.math.BigDecimal
 import org.springframework.context.annotation.Profile
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Slice
@@ -34,7 +35,7 @@ interface AppRoundActionSummaryRepository :
     fun findByIdAndRoundId(id: String, roundId: Int): AppRoundActionSummary?
 
     fun countByTotalRewardAmountGreaterThanAndAppIdAndRoundId(
-        totalRewardAmount: Double,
+        totalRewardAmount: BigDecimal,
         appId: String,
         roundId: Int,
     ): Long
