+++ packages/common/src/main/kotlin/org/vechain/indexer/b3tr/action/repository/UserAllTimeActionSummaryRepository.kt
@@ -1,5 +1,6 @@
 package org.vechain.indexer.b3tr.action.repository
 
+import java.math.BigDecimal
 import org.springframework.context.annotation.Profile
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Slice
@@ -15,7 +16,7 @@ interface UserAllTimeActionSummaryRepository :
     // Count entries where totalRewardAmount is greater than a specific value, filtering by entity
     // type
     fun countByTotalRewardAmountGreaterThanAndEntityType(
-        totalRewardAmount: Double,
+        totalRewardAmount: BigDecimal,
         entityType: EntityType,
     ): Long
