+++ packages/common/src/main/kotlin/org/vechain/indexer/b3tr/shared/AppActionSummaryDocument.kt
@@ -1,5 +1,6 @@
 package org.vechain.indexer.b3tr.shared
 
+import java.math.BigDecimal
 import org.vechain.indexer.VersionedDocument
 import org.vechain.indexer.b3tr.action.Impact
 
@@ -7,6 +8,6 @@ interface AppActionSummaryDocument : VersionedDocument {
     val appId: String
     val user: String
     val actionsRewarded: Long
-    val totalRewardAmount: Double
+    val totalRewardAmount: BigDecimal
     val totalImpact: Impact?
 }
